警告対処

+++ Source/Plugins/DirectX12RHI/Private/Plugins/DirectX12RHI/Descriptor/DescriptorHeap.cpp
@@ -183,7 +183,7 @@ namespace ob::rhi::dx12 {
 		OB_ASSERT_RANGE(index, 0, m_capacity-1);
 		OB_ASSERT(m_heap,"ヒープが空です。");
 		D3D12_CPU_DESCRIPTOR_HANDLE handle = m_heap->GetCPUDescriptorHandleForHeapStart();
-		handle.ptr += index * m_descriptorSize;
+		handle.ptr += static_cast<size_t>(index) * m_descriptorSize;
 		return handle;
 	}
 
@@ -195,7 +195,7 @@ namespace ob::rhi::dx12 {
 		OB_ASSERT_RANGE(index, 0, m_capacity-1);
 		OB_ASSERT(m_heap, "ヒープが空です。");
 		D3D12_GPU_DESCRIPTOR_HANDLE handle = m_heap->GetGPUDescriptorHandleForHeapStart();
-		handle.ptr += index * m_descriptorSize;
+		handle.ptr += static_cast<size_t>(index) * m_descriptorSize;
 		return handle;
 	}
